+++ src/app/(main)/boards/knowhow/_components/KnowhowFilter.tsx
@@ -1,9 +1,9 @@
 "use client";
-import { FormEventHandler, useRef } from "react";
-import { SEARCH_OPTIONS, SORT_OPTIONS, TOption } from "../_constants";
+import { TOption } from "../_constants";
+import SearchOptions from "./SearchOptions";
+import SortOptions from "./SortOptions";
 
 type KnowhowFilterTypes = {
-  sortOrder: TOption["value"];
   selectedSearchOption: TOption["value"];
   onSortOrderChange: (value: TOption["value"]) => void;
   onSearchOptionChange: (value: TOption["value"]) => void;
@@ -11,45 +11,20 @@ type KnowhowFilterTypes = {
 };
 
 function KnowhowFilter({
-  sortOrder,
   selectedSearchOption,
   onSortOrderChange,
   onSearchOptionChange,
   onSearch
 }: KnowhowFilterTypes) {
-  const searchInputRef = useRef<HTMLInputElement>(null);
-  const handleSearch: FormEventHandler<HTMLFormElement> = (e) => {
-    e.preventDefault();
-    if (searchInputRef.current) {
-      onSearch(searchInputRef.current.value);
-    }
-  };
-
   return (
-    <div className=" flex">
-      <div>
-        {SORT_OPTIONS.map((option) => (
-          <button key={option.value} onClick={() => onSortOrderChange(option.value)}>
-            {option.label}
-          </button>
-        ))}
-      </div>
-      <div className="flex">
-        <div>
-          <select value={selectedSearchOption} onChange={(e) => onSearchOptionChange(e.target.value)}>
-            {SEARCH_OPTIONS.map((option) => (
-              <option key={option.value} value={option.value}>
-                {option.label}
-              </option>
-            ))}
-          </select>
-        </div>
-        <form onSubmit={handleSearch}>
-          <input ref={searchInputRef} type="text" placeholder="검색어를 입력해주세요" />
-          <button>돋보기</button>
-        </form>
-      </div>
-    </div>
+    <nav className="flex">
+      <SortOptions onSortOrderChange={onSortOrderChange} />
+      <SearchOptions
+        onSearch={onSearch}
+        onSearchOptionChange={onSearchOptionChange}
+        selectedSearchOption={selectedSearchOption}
+      />
+    </nav>
   );
 }
